Add option to set a bigger column then others
The problem was the middle column had more content then the others, so
if it had a height set on it at small screen sizes the content would go
outside of the container.

This creates a "grow" class which we can put on this column so it's
height is auto and still allows the other columns to all line up.